see if moving into subdir makes build pass
