+++ packages/patternfly-4/react-core/src/components/withOuia/index.ts
@@ -1,0 +1,1 @@
+export * from './withOuia';
