+++ packages/patternfly-4/react-core/src/components/withOuia/withOuia.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { isOUIAEnvironment, getUniqueId, generateOUIAId } from '../helpers/ouia';
+import { isOUIAEnvironment, getUniqueId, generateOUIAId } from './ouia';
 
 export const OuiaContext = React.createContext<OuiaContextProps | null>(null);
